Port RDMA Communication Manager

+++ src/main/java/de/hhu/bsinfo/neutrino/data/NativeByteBitMask.java
@@ -2,9 +2,10 @@ package de.hhu.bsinfo.neutrino.data;
 
 import de.hhu.bsinfo.neutrino.buffer.LocalBuffer;
 import de.hhu.bsinfo.neutrino.util.BitMask;
+import de.hhu.bsinfo.neutrino.util.flag.ByteFlag;
 import de.hhu.bsinfo.neutrino.util.flag.LongFlag;
 
-public class NativeByteBitMask<T extends Enum<T> & LongFlag> extends NativeDataType {
+public class NativeByteBitMask<T extends Enum<T> & ByteFlag> extends NativeDataType {
 
     public NativeByteBitMask(LocalBuffer byteBuffer, final long offset) {
         super(byteBuffer, offset);
