+++ src/main/java/de/hhu/bsinfo/neutrino/data/NativeIntegerBitMask.java
@@ -2,9 +2,10 @@ package de.hhu.bsinfo.neutrino.data;
 
 import de.hhu.bsinfo.neutrino.buffer.LocalBuffer;
 import de.hhu.bsinfo.neutrino.util.BitMask;
+import de.hhu.bsinfo.neutrino.util.flag.IntegerFlag;
 import de.hhu.bsinfo.neutrino.util.flag.LongFlag;
 
-public class NativeIntegerBitMask<T extends Enum<T> & LongFlag> extends NativeDataType {
+public class NativeIntegerBitMask<T extends Enum<T> & IntegerFlag> extends NativeDataType {
 
     public NativeIntegerBitMask(LocalBuffer byteBuffer, final long offset) {
         super(byteBuffer, offset);
