+++ src/main/java/de/hhu/bsinfo/neutrino/util/Epoll.java
@@ -4,6 +4,7 @@ import de.hhu.bsinfo.neutrino.data.NativeArray;
 import de.hhu.bsinfo.neutrino.data.NativeIntegerBitMask;
 import de.hhu.bsinfo.neutrino.data.NativeLong;
 import de.hhu.bsinfo.neutrino.struct.Struct;
+import de.hhu.bsinfo.neutrino.util.flag.IntegerFlag;
 import de.hhu.bsinfo.neutrino.util.flag.LongFlag;
 
 public final class Epoll {
@@ -73,7 +74,7 @@ public final class Epoll {
         }
     }
 
-    public enum EventType implements LongFlag {
+    public enum EventType implements IntegerFlag {
         EPOLLIN(0x001), EPOLLPRI(0x002), EPOLLOUT(0x004), EPOLLRDNORM(0x040), EPOLLRDBAND(0x080), EPOLLWRNORM(0x100),
         EPOLLWRBAND(0x200), EPOLLMSG(0x400), EPOLLERR(0x008), EPOLLHUP(0x010), EPOLLRDHUP(0x2000), EPOLLONESHOT(1 << 30),
         EPOLLET(1 << 31);
@@ -85,7 +86,7 @@ public final class Epoll {
         }
 
         @Override
-        public long getValue() {
+        public int getValue() {
             return value;
         }
     }
