+++ src/main/java/de/hhu/bsinfo/neutrino/verbs/SendWorkRequest.java
@@ -4,6 +4,7 @@ import de.hhu.bsinfo.neutrino.buffer.LocalBuffer;
 import de.hhu.bsinfo.neutrino.data.*;
 import de.hhu.bsinfo.neutrino.struct.Struct;
 import de.hhu.bsinfo.neutrino.util.*;
+import de.hhu.bsinfo.neutrino.util.flag.IntegerFlag;
 import de.hhu.bsinfo.neutrino.util.flag.LongFlag;
 
 import java.util.Arrays;
@@ -417,7 +418,7 @@ public class SendWorkRequest extends Struct implements Linkable<SendWorkRequest>
         };
     }
 
-    public enum SendFlag implements LongFlag {
+    public enum SendFlag implements IntegerFlag {
         FENCE(1), SIGNALED(1 << 1), SOLICITED(1 << 2), INLINE(1 << 3), IP_CSUM(1 << 4);
 
         private final int value;
@@ -427,7 +428,7 @@ public class SendWorkRequest extends Struct implements Linkable<SendWorkRequest>
         }
 
         @Override
-        public long getValue() {
+        public int getValue() {
             return value;
         }
     }
