+++ src/main/java/de/hhu/bsinfo/neutrino/verbs/WorkQueue.java
@@ -4,6 +4,7 @@ import de.hhu.bsinfo.neutrino.buffer.LocalBuffer;
 import de.hhu.bsinfo.neutrino.data.*;
 import de.hhu.bsinfo.neutrino.struct.Result;
 import de.hhu.bsinfo.neutrino.struct.Struct;
+import de.hhu.bsinfo.neutrino.util.flag.IntegerFlag;
 import de.hhu.bsinfo.neutrino.util.flag.LongFlag;
 import de.hhu.bsinfo.neutrino.util.LinkNative;
 import de.hhu.bsinfo.neutrino.util.NativeObjectRegistry;
@@ -221,7 +222,7 @@ public class WorkQueue extends Struct implements AutoCloseable {
         };
     }
     
-    public enum AttributeFlag implements LongFlag {
+    public enum AttributeFlag implements IntegerFlag {
         STATE(1), CURRENT_STATE(1 << 1), FLAGS(1 << 2), RESERVED(1 << 3);
 
         private final int value;
@@ -231,12 +232,12 @@ public class WorkQueue extends Struct implements AutoCloseable {
         }
 
         @Override
-        public long getValue() {
+        public int getValue() {
             return value;
         }
     }
 
-    public enum CompatibilityFlag implements LongFlag {
+    public enum CompatibilityFlag implements IntegerFlag {
         FLAGS(1), RESERVED(1 << 1);
 
         private final int value;
@@ -246,12 +247,12 @@ public class WorkQueue extends Struct implements AutoCloseable {
         }
 
         @Override
-        public long getValue() {
+        public int getValue() {
             return value;
         }
     }
 
-    public enum CreationFlag implements LongFlag {
+    public enum CreationFlag implements IntegerFlag {
         CVLAN_STRIPPING(1), SCATTER_FCS(1 << 1), DELAY_DROP(1 << 2),
         PCI_WRITE_END_PADDING(1 << 3), RESERVED(1 << 4);
 
@@ -262,7 +263,7 @@ public class WorkQueue extends Struct implements AutoCloseable {
         }
 
         @Override
-        public long getValue() {
+        public int getValue() {
             return value;
         }
     }
